shuheng

+++ src/main/java/cn/stylefeng/guns/modular/system/controller/UserMgrController.java
@@ -24,12 +24,16 @@ import cn.stylefeng.guns.core.common.constant.dictmap.UserDict;
 import cn.stylefeng.guns.core.common.constant.factory.ConstantFactory;
 import cn.stylefeng.guns.core.common.constant.state.ManagerStatus;
 import cn.stylefeng.guns.core.common.exception.BizExceptionEnum;
+import cn.stylefeng.guns.core.common.node.ZTreeNode;
 import cn.stylefeng.guns.core.common.page.LayuiPageFactory;
 import cn.stylefeng.guns.core.log.LogObjectHolder;
 import cn.stylefeng.guns.core.shiro.ShiroKit;
+import cn.stylefeng.guns.core.shiro.ShiroUser;
+import cn.stylefeng.guns.modular.system.entity.Role;
 import cn.stylefeng.guns.modular.system.entity.User;
 import cn.stylefeng.guns.modular.system.factory.UserFactory;
 import cn.stylefeng.guns.modular.system.model.UserDto;
+import cn.stylefeng.guns.modular.system.service.RoleService;
 import cn.stylefeng.guns.modular.system.service.UserService;
 import cn.stylefeng.guns.modular.system.warpper.UserWrapper;
 import cn.stylefeng.roses.core.base.controller.BaseController;
@@ -38,6 +42,7 @@ import cn.stylefeng.roses.core.reqres.response.ResponseData;
 import cn.stylefeng.roses.core.util.ToolUtil;
 import cn.stylefeng.roses.kernel.model.exception.RequestEmptyException;
 import cn.stylefeng.roses.kernel.model.exception.ServiceException;
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -48,7 +53,9 @@ import org.springframework.web.multipart.MultipartFile;
 
 import javax.validation.Valid;
 import java.io.File;
+import java.sql.Wrapper;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 
@@ -69,6 +76,9 @@ public class UserMgrController extends BaseController {
 
     @Autowired
     private UserService userService;
+
+    @Autowired
+    private RoleService roleService;
 
     /**
      * 跳转到查看管理员列表的页面
@@ -177,7 +187,7 @@ public class UserMgrController extends BaseController {
     @ResponseBody
     public Object list(@RequestParam(required = false) String name,
                        @RequestParam(required = false) String timeLimit,
-                       @RequestParam(required = false) Long deptId) {
+                       @RequestParam(required = false) Long plazaId) {
 
         //拼接查询条件
         String beginTime = "";
@@ -190,14 +200,18 @@ public class UserMgrController extends BaseController {
         }
 
         if (ShiroKit.isAdmin()) {
-            Page<Map<String, Object>> users = userService.selectUsers(null, name, beginTime, endTime, deptId);
+            Page<Map<String, Object>> users = userService.selectUsers(null, name, beginTime, endTime, plazaId);
             Page wrapped = new UserWrapper(users).wrap();
             return LayuiPageFactory.createPageInfo(wrapped);
-        } else {
-            DataScope dataScope = new DataScope(ShiroKit.getDeptDataScope());
-            Page<Map<String, Object>> users = userService.selectUsers(dataScope, name, beginTime, endTime, deptId);
+        } else if(ShiroKit.isGeneral() && ShiroKit.isPlazaAdmin()){
+            //纪念馆管理员
+            plazaId = ShiroKit.getUser().getPlazaId();
+            Page<Map<String, Object>> users = userService.selectUsers(null, name, beginTime, endTime, plazaId);
             Page wrapped = new UserWrapper(users).wrap();
             return LayuiPageFactory.createPageInfo(wrapped);
+        }else{
+            Page<Map<String, Object>> page = new Page<>();
+            return LayuiPageFactory.createPageInfo(page);
         }
     }
 
@@ -209,11 +223,23 @@ public class UserMgrController extends BaseController {
      */
     @RequestMapping("/add")
     @BussinessLog(value = "添加管理员", key = "account", dict = UserDict.class)
-    @Permission(Const.ADMIN_NAME)
+    @Permission({Const.ADMIN_NAME,Const.GENERAL_NAME})
     @ResponseBody
-    public ResponseData add(@Valid UserDto user, BindingResult result) {
-        if (result.hasErrors()) {
+    public ResponseData add(UserDto user) {
+        if (ToolUtil.isOneEmpty(user,user.getName(),user.getPhone(),user.getAccount(),user.getPassword(),user.getSex())) {
             throw new ServiceException(BizExceptionEnum.REQUEST_NULL);
+        }
+        //普通管理员添加用户只能是自己纪念馆的
+        if(ShiroKit.isGeneral()){
+            user.setPlazaId(ShiroKit.getUser().getPlazaId());
+            //设置role_id
+            QueryWrapper wrapper = new QueryWrapper<Role>();
+            wrapper.eq("description",Const.GENERAL_NAME);
+            Role role = roleService.getOne(wrapper);
+            if(ToolUtil.isOneEmpty(role,role.getRoleId())){
+                throw new ServiceException(BizExceptionEnum.REQUEST_NULL);
+            }
+            user.setRoleId(role.getRoleId().toString());
         }
         this.userService.addUser(user);
         return SUCCESS_TIP;
@@ -228,9 +254,14 @@ public class UserMgrController extends BaseController {
     @RequestMapping("/edit")
     @BussinessLog(value = "修改管理员", key = "account", dict = UserDict.class)
     @ResponseBody
-    public ResponseData edit(@Valid UserDto user, BindingResult result) {
-        if (result.hasErrors()) {
+    public ResponseData edit(UserDto user) {
+        if (ToolUtil.isOneEmpty(user,user.getUserId(),user.getName(),user.getPhone(),user.getSex())) {
             throw new ServiceException(BizExceptionEnum.REQUEST_NULL);
+        }
+        ShiroUser shiroUser = ShiroKit.getUser();
+        if(ShiroKit.isGeneral() && !shiroUser.isPlazaAdmin()){
+            //纪念馆管理员没有总权限
+            throw new ServiceException(BizExceptionEnum.NO_PERMITION);
         }
         this.userService.editUser(user);
         return SUCCESS_TIP;
@@ -249,6 +280,14 @@ public class UserMgrController extends BaseController {
     public ResponseData delete(@RequestParam Long userId) {
         if (ToolUtil.isEmpty(userId)) {
             throw new ServiceException(BizExceptionEnum.REQUEST_NULL);
+        }
+        ShiroUser user = ShiroKit.getUser();
+        if(user.getId().equals(userId)){
+            throw new ServiceException(BizExceptionEnum.LIMIT_OPERATE_SELF);
+        }
+        if(ShiroKit.isGeneral() && !user.isPlazaAdmin()){
+            //纪念馆管理员没有总权限
+            throw new ServiceException(BizExceptionEnum.NO_PERMITION);
         }
         this.userService.deleteUser(userId);
         return SUCCESS_TIP;
@@ -278,11 +317,16 @@ public class UserMgrController extends BaseController {
      */
     @RequestMapping("/reset")
     @BussinessLog(value = "重置管理员密码", key = "userId", dict = UserDict.class)
-    @Permission(Const.ADMIN_NAME)
+    @Permission({Const.ADMIN_NAME,Const.GENERAL_NAME})
     @ResponseBody
     public ResponseData reset(@RequestParam Long userId) {
         if (ToolUtil.isEmpty(userId)) {
             throw new ServiceException(BizExceptionEnum.REQUEST_NULL);
+        }
+        ShiroUser shiroUser = ShiroKit.getUser();
+        if(ShiroKit.isGeneral() && !shiroUser.isPlazaAdmin()){
+            //纪念馆管理员没有总权限
+            throw new ServiceException(BizExceptionEnum.NO_PERMITION);
         }
         this.userService.assertAuth(userId);
         User user = this.userService.getById(userId);
@@ -300,7 +344,7 @@ public class UserMgrController extends BaseController {
      */
     @RequestMapping("/freeze")
     @BussinessLog(value = "冻结用户", key = "userId", dict = UserDict.class)
-    @Permission(Const.ADMIN_NAME)
+    @Permission({Const.ADMIN_NAME,Const.GENERAL_NAME})
     @ResponseBody
     public ResponseData freeze(@RequestParam Long userId) {
         if (ToolUtil.isEmpty(userId)) {
@@ -310,6 +354,18 @@ public class UserMgrController extends BaseController {
         if (userId.equals(Const.ADMIN_ID)) {
             throw new ServiceException(BizExceptionEnum.CANT_FREEZE_ADMIN);
         }
+
+        ShiroUser user = ShiroKit.getUser();
+        //不能冻结自己
+        if(user.getId().equals(userId)){
+            throw new ServiceException(BizExceptionEnum.LIMIT_OPERATE_SELF);
+        }
+
+        if(ShiroKit.isGeneral() && !user.isPlazaAdmin()){
+            //纪念馆管理员没有总权限
+            throw new ServiceException(BizExceptionEnum.NO_PERMITION);
+        }
+
         this.userService.assertAuth(userId);
         this.userService.setStatus(userId, ManagerStatus.FREEZED.getCode());
         return SUCCESS_TIP;
@@ -323,11 +379,20 @@ public class UserMgrController extends BaseController {
      */
     @RequestMapping("/unfreeze")
     @BussinessLog(value = "解除冻结用户", key = "userId", dict = UserDict.class)
-    @Permission(Const.ADMIN_NAME)
+    @Permission({Const.ADMIN_NAME,Const.GENERAL_NAME})
     @ResponseBody
     public ResponseData unfreeze(@RequestParam Long userId) {
         if (ToolUtil.isEmpty(userId)) {
             throw new ServiceException(BizExceptionEnum.REQUEST_NULL);
+        }
+        ShiroUser user = ShiroKit.getUser();
+        if(user.getId().equals(userId)){
+            throw new ServiceException(BizExceptionEnum.LIMIT_OPERATE_SELF);
+        }
+
+        if(ShiroKit.isGeneral() && !user.isPlazaAdmin()){
+            //纪念馆管理员没有总权限
+            throw new ServiceException(BizExceptionEnum.NO_PERMITION);
         }
         this.userService.assertAuth(userId);
         this.userService.setStatus(userId, ManagerStatus.OK.getCode());
@@ -344,7 +409,7 @@ public class UserMgrController extends BaseController {
     @BussinessLog(value = "分配角色", key = "userId,roleIds", dict = UserDict.class)
     @Permission(Const.ADMIN_NAME)
     @ResponseBody
-    public ResponseData setRole(@RequestParam("userId") Long userId, @RequestParam("roleIds") String roleIds) {
+    public ResponseData setRole(@RequestParam("userId") Long userId, @RequestParam("roleIds") String roleIds, @RequestParam(value = "plazaId",required = false) Long plazaId) {
         if (ToolUtil.isOneEmpty(userId, roleIds)) {
             throw new ServiceException(BizExceptionEnum.REQUEST_NULL);
         }
@@ -354,6 +419,76 @@ public class UserMgrController extends BaseController {
         }
         this.userService.assertAuth(userId);
         this.userService.setRoles(userId, roleIds);
+        //保存关联的纪念馆id,设置纪念馆管理员
+        User user = userService.getById(userId);
+        if(ToolUtil.isNotEmpty(plazaId)){
+            user.setPlazaAdmin(true);
+            user.setPlazaId(plazaId);
+            //删除sys_user_menu
+            userService.deleteMenuIdsByUserId(userId);
+        }else{
+            user.setPlazaAdmin(false);
+            user.setPlazaId(null);
+        }
+        this.userService.updateById(user);
+        return SUCCESS_TIP;
+    }
+
+
+
+    /**
+     * 跳转到权限分配页面
+     *
+     * @author fengshuonan
+     * @Date 2018/12/24 22:43
+     */
+    @Permission
+    @RequestMapping("/permission_assign")
+    public String permissionAssign(@RequestParam Long userId, Model model) {
+        if (ToolUtil.isEmpty(userId)) {
+            throw new ServiceException(BizExceptionEnum.REQUEST_NULL);
+        }
+        model.addAttribute("userId", userId);
+        return PREFIX + "permission_roleassign.html";
+    }
+
+
+    @RequestMapping("/setPermission")
+    @BussinessLog(value = "分配权限", key = "userId,menuIds", dict = UserDict.class)
+    @Permission({Const.ADMIN_NAME,Const.GENERAL_NAME})
+    @ResponseBody
+    public ResponseData setPermission(@RequestParam("userId") Long userId, @RequestParam(value = "menuIds",required = false) String menuIds, @RequestParam(value = "isPlazaAdmin") Integer isPlazaAdmin){
+        ShiroUser shiroUser = ShiroKit.getUser();
+        if (ToolUtil.isOneEmpty(userId,isPlazaAdmin)) {
+            throw new ServiceException(BizExceptionEnum.REQUEST_NULL);
+        }
+        if(userId.equals(shiroUser.getId())){
+            //不允许修改自己权限
+            throw new ServiceException(BizExceptionEnum.LIMIT_OPERATE_SELF);
+        }
+        if(ShiroKit.isGeneral() && !shiroUser.isPlazaAdmin()){
+            //纪念馆管理员但没有总权限
+            throw new ServiceException(BizExceptionEnum.NO_PERMITION);
+        }
+        User user = this.userService.getById(userId);
+        user.setPlazaId(shiroUser.getPlazaId());
+        if(isPlazaAdmin == 1){
+            //是场馆管理员
+            QueryWrapper wrapper = new QueryWrapper<Role>();
+            wrapper.eq("description",Const.GENERAL_NAME);
+            Role role = roleService.getOne(wrapper);
+            if(ToolUtil.isOneEmpty(role,role.getRoleId())){
+                throw new ServiceException(BizExceptionEnum.REQUEST_NULL);
+            }
+            user.setRoleId(role.getRoleId().toString());
+            user.setPlazaAdmin(true);
+            userService.updateById(user);
+        }
+        if(isPlazaAdmin == 0){
+            //自定义权限
+            user.setPlazaAdmin(false);
+            userService.setPermission(user,menuIds);
+        }
         return SUCCESS_TIP;
     }
 
@@ -375,5 +510,22 @@ public class UserMgrController extends BaseController {
             throw new ServiceException(BizExceptionEnum.UPLOAD_ERROR);
         }
         return pictureName;
+    }
+
+    /**
+     * 根据用户id查询拥有权限的普通管理员菜单树
+     * @param userId
+     * @return
+     */
+    @RequestMapping(value = "/menuTreeByUserId/{userId}")
+    @ResponseBody
+    public List<ZTreeNode> menuTreeByUserId(@PathVariable Long userId) {
+        List<Long> menuIds = this.userService.getMenuIdsByUserId(userId);
+
+        if (ToolUtil.isEmpty(menuIds)) {
+            return this.userService.menuTreeList(Const.GENERAL_NAME);
+        } else {
+            return this.userService.menuTreeListByMenuIds(Const.GENERAL_NAME,menuIds);
+        }
     }
 }
